+++ extensions/markdown-language-features/package.json
@@ -604,7 +604,6 @@
     "morphdom": "^2.6.1",
     "picomatch": "^2.3.1",
     "vscode-languageclient": "^8.0.2",
-    "vscode-languageserver-textdocument": "^1.0.4",
     "vscode-nls": "^5.1.0",
     "vscode-uri": "^3.0.3"
   },
