+++ extensions/markdown-language-features/src/extension.ts
@@ -4,14 +4,13 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import { BaseLanguageClient, LanguageClient, ServerOptions, TransportKind } from 'vscode-languageclient/node';
-import { startClient } from './client';
+import { LanguageClient, ServerOptions, TransportKind } from 'vscode-languageclient/node';
+import { MdLanguageClient, startClient } from './client/client';
 import { activateShared } from './extension.shared';
 import { VsCodeOutputLogger } from './logging';
 import { IMdParser, MarkdownItEngine } from './markdownEngine';
 import { getMarkdownExtensionContributions } from './markdownExtensions';
 import { githubSlugifier } from './slugify';
-import { IMdWorkspace, VsCodeMdWorkspace } from './workspace';
 
 export async function activate(context: vscode.ExtensionContext) {
 	const contributions = getMarkdownExtensionContributions(context);
@@ -22,17 +21,12 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	const engine = new MarkdownItEngine(contributions, githubSlugifier, logger);
 
-	const workspace = new VsCodeMdWorkspace();
-	context.subscriptions.push(workspace);
-
-	const client = await startServer(context, workspace, engine);
-	context.subscriptions.push({
-		dispose: () => client.stop()
-	});
+	const client = await startServer(context, engine);
+	context.subscriptions.push(client);
 	activateShared(context, client, engine, logger, contributions);
 }
 
-function startServer(context: vscode.ExtensionContext, workspace: IMdWorkspace, parser: IMdParser): Promise<BaseLanguageClient> {
+function startServer(context: vscode.ExtensionContext, parser: IMdParser): Promise<MdLanguageClient> {
 	const clientMain = vscode.extensions.getExtension('vscode.markdown-language-features')?.packageJSON?.main || '';
 
 	const serverMain = `./server/${clientMain.indexOf('/dist/') !== -1 ? 'dist' : 'out'}/node/main`;
@@ -49,5 +43,5 @@ function startServer(context: vscode.ExtensionContext, workspace: IMdWorkspace, 
 	};
 	return startClient((id, name, clientOptions) => {
 		return new LanguageClient(id, name, serverOptions, clientOptions);
-	}, workspace, parser);
+	}, parser);
 }
