+++ extensions/markdown-language-features/src/languageFeatures/fileReferences.ts
@@ -4,11 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import { BaseLanguageClient } from 'vscode-languageclient';
 import type * as lsp from 'vscode-languageserver-types';
 import * as nls from 'vscode-nls';
+import { MdLanguageClient } from '../client/client';
 import { Command, CommandManager } from '../commandManager';
-import { getReferencesToFileInWorkspace } from '../protocol';
 
 const localize = nls.loadMessageBundle();
 
@@ -18,7 +17,7 @@ export class FindFileReferencesCommand implements Command {
 	public readonly id = 'markdown.findAllFileReferences';
 
 	constructor(
-		private readonly client: BaseLanguageClient,
+		private readonly client: MdLanguageClient,
 	) { }
 
 	public async execute(resource?: vscode.Uri) {
@@ -35,7 +34,7 @@ export class FindFileReferencesCommand implements Command {
 			location: vscode.ProgressLocation.Window,
 			title: localize('progress.title', "Finding file references")
 		}, async (_progress, token) => {
-			const locations = (await this.client.sendRequest(getReferencesToFileInWorkspace, { uri: resource!.toString() }, token)).map(loc => {
+			const locations = (await this.client.getReferencesToFileInWorkspace(resource!, token)).map(loc => {
 				return new vscode.Location(vscode.Uri.parse(loc.uri), convertRange(loc.range));
 			});
 
@@ -58,7 +57,7 @@ export function convertRange(range: lsp.Range): vscode.Range {
 
 export function registerFindFileReferenceSupport(
 	commandManager: CommandManager,
-	client: BaseLanguageClient,
+	client: MdLanguageClient,
 ): vscode.Disposable {
 	return commandManager.register(new FindFileReferencesCommand(client));
 }
