+++ extensions/markdown-language-features/src/types/textDocument.ts
@@ -11,12 +11,7 @@ import * as vscode from 'vscode';
 export interface ITextDocument {
 	readonly uri: vscode.Uri;
 	readonly version: number;
-	readonly lineCount: number;
 
-	getText(range?: vscode.Range): string;
-	positionAt(offset: number): vscode.Position;
+	getText(): string;
 }
 
-export function getLine(doc: ITextDocument, line: number): string {
-	return doc.getText(new vscode.Range(line, 0, line, Number.MAX_VALUE)).replace(/\r?\n$/, '');
-}
